shifted return value of n_choose_k fucntion to uint

+++ src/utils.cpp
@@ -4,8 +4,8 @@
 namespace bgen {
 
 // Returns value of Binomial Coefficient C(n, k)
-int n_choose_k(int n, int k) {
-  int res = 1;
+uint n_choose_k(int n, int k) {
+  uint res = 1;
 
   // Since C(n, k) = C(n, n-k)
   if ( k > n - k ) {
@@ -14,7 +14,7 @@ int n_choose_k(int n, int k) {
 
   // Calculate value of
   // [n * (n-1) *---* (n-k+1)] / [k * (k-1) *----* 1]
-  for (int i = 0; i < k; ++i) {
+  for (uint i = 0; i < (uint) k; ++i) {
     res *= (n - i);
     res /= (i + 1);
   }
